+++ src/components/Shared/PPM/Closeout/AboutForm/AboutForm.jsx
@@ -8,6 +8,7 @@ import * as Yup from 'yup';
 import styles from './AboutForm.module.scss';
 
 import ppmStyles from 'components/Shared/PPM/PPM.module.scss';
+import closingPageStyles from 'components/Shared/PPM/Closeout/Closeout.module.scss';
 import SectionWrapper from 'components/Customer/SectionWrapper';
 import { DatePickerInput } from 'components/form/fields';
 import MaskedTextField from 'components/form/fields/MaskedTextField/MaskedTextField';
@@ -82,199 +83,228 @@ const AboutForm = ({ mtoShipment, onBack, onSubmit, isSubmitted, appName }) => {
   };
 
   return (
-    <Formik initialValues={initialValues} validationSchema={validationSchema} onSubmit={onSubmit}>
-      {({ isValid, isSubmitting, handleSubmit, values, ...formikProps }) => {
-        return (
-          <div className={classnames(ppmStyles.formContainer, styles.AboutForm)}>
-            <Form className={classnames(formStyles.form, ppmStyles.form, styles.W2Address)} data-testid="aboutForm">
-              <SectionWrapper className={classnames(ppmStyles.sectionWrapper, formStyles.formSection)}>
-                <h2>Departure date</h2>
-                <DatePickerInput
-                  disabledDays={{ after: today }}
-                  className={classnames(styles.actualMoveDate, 'usa-input')}
-                  name="actualMoveDate"
-                  label="When did you leave your origin?"
-                />
-                <Hint className={ppmStyles.hint}>If it took you more than one day to move out, use the first day.</Hint>
-                <h2>Locations</h2>
-                <p>
-                  If you picked things up or dropped things off from other places a long way from your start or end
-                  ZIPs, ask your counselor if you should add another PPM shipment.
-                </p>
-                <AddressFields
-                  name="pickupAddress"
-                  legend="Pickup Address"
-                  labelHint="Required"
-                  locationLookup
-                  formikProps={formikProps}
-                  className={styles.AddressFieldSet}
-                  render={(fields) => (
-                    <>
-                      {fields}
-                      <h4>Second Pickup Address</h4>
-                      <FormGroup>
-                        <p>
-                          Will you pick up any belongings from a second address? (Must be near the pickup address.
-                          Subject to approval.)
-                        </p>
-                        <div className={formStyles.radioGroup}>
-                          <Field
-                            as={Radio}
-                            id="has-secondary-pickup"
-                            data-testid="has-secondary-pickup"
-                            label="Yes"
-                            name="hasSecondaryPickupAddress"
-                            value="true"
-                            title="Yes, there is a second pickup address"
-                            checked={values.hasSecondaryPickupAddress === 'true'}
+    <>
+      <div className={classnames(closingPageStyles['closing-section'], closingPageStyles['about-ppm'])}>
+        <p>Finish moving this PPM before you start documenting it.</p>
+        <h2>How to complete your PPM</h2>
+        <p>To complete your PPM, you will:</p>
+        <ul>
+          <li>Upload weight tickets for each trip</li>
+          <li>Upload receipts to document any expenses</li>
+          <li>Upload receipts if you used short-term storage, so you can request reimbursement</li>
+          <li>Upload any other documentation (such as proof of ownership for a trailer, if you used your own)</li>
+          <li>Complete your PPM to send it to a counselor for review</li>
+        </ul>
+        <h2>About your final payment</h2>
+        <p>Your final payment will be:</p>
+        <ul>
+          <li>based on your final incentive</li>
+          <li>modified by expenses submitted (authorized expenses reduce your tax burden)</li>
+          <li>minus any taxes withheld (the IRS considers your incentive to be taxable income)</li>
+          <li>plus any reimbursements you receive</li>
+        </ul>
+        <p>
+          Verified expenses reduce the taxable income you report to the IRS on form W-2. They may not be claimed again
+          as moving expenses. Federal tax withholding will be deducted from the profit (entitlement less eligible
+          operating expenses.)
+        </p>
+      </div>
+      <Formik initialValues={initialValues} validationSchema={validationSchema} onSubmit={onSubmit}>
+        {({ isValid, isSubmitting, handleSubmit, values, ...formikProps }) => {
+          return (
+            <div className={classnames(ppmStyles.formContainer, styles.AboutForm)}>
+              <Form className={classnames(formStyles.form, ppmStyles.form, styles.W2Address)} data-testid="aboutForm">
+                <SectionWrapper className={classnames(ppmStyles.sectionWrapper, formStyles.formSection)}>
+                  <h2>Departure date</h2>
+                  <DatePickerInput
+                    disabledDays={{ after: today }}
+                    className={classnames(styles.actualMoveDate, 'usa-input')}
+                    name="actualMoveDate"
+                    label="When did you leave your origin?"
+                  />
+                  <Hint className={ppmStyles.hint}>
+                    If it took you more than one day to move out, use the first day.
+                  </Hint>
+                  <h2>Locations</h2>
+                  <p>
+                    If you picked things up or dropped things off from other places a long way from your start or end
+                    ZIPs, ask your counselor if you should add another PPM shipment.
+                  </p>
+                  <AddressFields
+                    name="pickupAddress"
+                    legend="Pickup Address"
+                    labelHint="Required"
+                    locationLookup
+                    formikProps={formikProps}
+                    className={styles.AddressFieldSet}
+                    render={(fields) => (
+                      <>
+                        {fields}
+                        <h4>Second Pickup Address</h4>
+                        <FormGroup>
+                          <p>
+                            Will you pick up any belongings from a second address? (Must be near the pickup address.
+                            Subject to approval.)
+                          </p>
+                          <div className={formStyles.radioGroup}>
+                            <Field
+                              as={Radio}
+                              id="has-secondary-pickup"
+                              data-testid="has-secondary-pickup"
+                              label="Yes"
+                              name="hasSecondaryPickupAddress"
+                              value="true"
+                              title="Yes, there is a second pickup address"
+                              checked={values.hasSecondaryPickupAddress === 'true'}
+                            />
+                            <Field
+                              as={Radio}
+                              id="no-secondary-pickup"
+                              data-testid="no-secondary-pickup"
+                              label="No"
+                              name="hasSecondaryPickupAddress"
+                              value="false"
+                              title="No, there is not a second pickup address"
+                              checked={values.hasSecondaryPickupAddress !== 'true'}
+                            />
+                          </div>
+                        </FormGroup>
+                        {values.hasSecondaryPickupAddress === 'true' && (
+                          <AddressFields
+                            name="secondaryPickupAddress"
+                            labelHint="Required"
+                            locationLookup
+                            formikProps={formikProps}
                           />
-                          <Field
-                            as={Radio}
-                            id="no-secondary-pickup"
-                            data-testid="no-secondary-pickup"
-                            label="No"
-                            name="hasSecondaryPickupAddress"
-                            value="false"
-                            title="No, there is not a second pickup address"
-                            checked={values.hasSecondaryPickupAddress !== 'true'}
-                          />
-                        </div>
-                      </FormGroup>
-                      {values.hasSecondaryPickupAddress === 'true' && (
-                        <AddressFields
-                          name="secondaryPickupAddress"
-                          labelHint="Required"
-                          locationLookup
-                          formikProps={formikProps}
-                        />
-                      )}
-                    </>
-                  )}
-                />
-                <AddressFields
-                  name="destinationAddress"
-                  legend="Delivery Address"
-                  className={styles.AddressFieldSet}
-                  labelHint="Required"
-                  locationLookup
-                  formikProps={formikProps}
-                  render={(fields) => (
-                    <>
-                      {fields}
-                      <h4>Second Delivery Address</h4>
-                      <FormGroup>
-                        <p>
-                          Will you deliver any belongings to a second address? (Must be near the delivery address.
-                          Subject to approval.)
-                        </p>
-                        <div className={formStyles.radioGroup}>
-                          <Field
-                            as={Radio}
-                            data-testid="has-secondary-destination"
-                            id="has-secondary-destination"
-                            label="Yes"
-                            name="hasSecondaryDestinationAddress"
-                            value="true"
-                            title="Yes, there is a second delivery address"
-                            checked={values.hasSecondaryDestinationAddress === 'true'}
-                          />
-                          <Field
-                            as={Radio}
-                            data-testid="no-secondary-destination"
-                            id="no-secondary-destination"
-                            label="No"
-                            name="hasSecondaryDestinationAddress"
-                            value="false"
-                            title="No, there is not a second delivery address"
-                            checked={values.hasSecondaryDestinationAddress !== 'true'}
-                          />
-                        </div>
-                      </FormGroup>
-                      {values.hasSecondaryDestinationAddress === 'true' && (
-                        <AddressFields
-                          name="secondaryDestinationAddress"
-                          labelHint="Required"
-                          locationLookup
-                          formikProps={formikProps}
-                        />
-                      )}
-                    </>
-                  )}
-                />
-                <h2>Advance (AOA)</h2>
-                <FormGroup>
-                  <Fieldset className={styles.advanceFieldset}>
-                    <legend className="usa-label">Did you receive an advance for this PPM?</legend>
-                    <Field
-                      as={Radio}
-                      id="yes-has-received-advance"
-                      data-testid="yes-has-received-advance"
-                      label="Yes"
-                      name="hasReceivedAdvance"
-                      value="true"
-                      checked={values.hasReceivedAdvance === 'true'}
-                    />
-                    <Field
-                      as={Radio}
-                      id="no-has-received-advance"
-                      data-testid="no-has-received-advance"
-                      label="No"
-                      name="hasReceivedAdvance"
-                      value="false"
-                      checked={values.hasReceivedAdvance === 'false'}
-                    />
-                    <Hint className={ppmStyles.hint}>
-                      If you requested an advance but did not receive it, select No.
-                    </Hint>
-                    {values.hasReceivedAdvance === 'true' && (
-                      <MaskedTextField
-                        label="How much did you receive?"
-                        name="advanceAmountReceived"
-                        id="advanceAmountReceived"
-                        defaultValue="0"
-                        mask={Number}
-                        scale={0} // digits after point, 0 for integers
-                        signed={false} // disallow negative
-                        thousandsSeparator=","
-                        lazy={false} // immediate masking evaluation
-                        prefix="$"
-                        hintClassName={ppmStyles.innerHint}
-                      />
+                        )}
+                      </>
                     )}
-                  </Fieldset>
-                </FormGroup>
-                <h2>W-2 address</h2>
-                <p>What is the address on your W-2?</p>
-                <AddressFields
-                  name={formFieldsName}
-                  className={styles.AddressFieldSet}
-                  labelHint="Required"
-                  locationLookup
-                  formikProps={formikProps}
-                />
-              </SectionWrapper>
-              <div
-                className={`${
-                  isCustomerPage ? ppmStyles.buttonContainer : `${formStyles.formActions} ${ppmStyles.buttonGroup}`
-                }`}
-              >
-                <Button className={ppmStyles.backButton} type="button" onClick={onBack} secondary outline>
-                  {`${isCustomerPage ? 'Return To Homepage' : 'Cancel'}`}
-                </Button>
-                <Button
-                  className={ppmStyles.saveButton}
-                  type="button"
-                  onClick={handleSubmit}
-                  disabled={!isValid || isSubmitted || isSubmitting}
+                  />
+                  <AddressFields
+                    name="destinationAddress"
+                    legend="Delivery Address"
+                    className={styles.AddressFieldSet}
+                    labelHint="Required"
+                    locationLookup
+                    formikProps={formikProps}
+                    render={(fields) => (
+                      <>
+                        {fields}
+                        <h4>Second Delivery Address</h4>
+                        <FormGroup>
+                          <p>
+                            Will you deliver any belongings to a second address? (Must be near the delivery address.
+                            Subject to approval.)
+                          </p>
+                          <div className={formStyles.radioGroup}>
+                            <Field
+                              as={Radio}
+                              data-testid="has-secondary-destination"
+                              id="has-secondary-destination"
+                              label="Yes"
+                              name="hasSecondaryDestinationAddress"
+                              value="true"
+                              title="Yes, there is a second delivery address"
+                              checked={values.hasSecondaryDestinationAddress === 'true'}
+                            />
+                            <Field
+                              as={Radio}
+                              data-testid="no-secondary-destination"
+                              id="no-secondary-destination"
+                              label="No"
+                              name="hasSecondaryDestinationAddress"
+                              value="false"
+                              title="No, there is not a second delivery address"
+                              checked={values.hasSecondaryDestinationAddress !== 'true'}
+                            />
+                          </div>
+                        </FormGroup>
+                        {values.hasSecondaryDestinationAddress === 'true' && (
+                          <AddressFields
+                            name="secondaryDestinationAddress"
+                            labelHint="Required"
+                            locationLookup
+                            formikProps={formikProps}
+                          />
+                        )}
+                      </>
+                    )}
+                  />
+                  <h2>Advance (AOA)</h2>
+                  <FormGroup>
+                    <Fieldset className={styles.advanceFieldset}>
+                      <legend className="usa-label">Did you receive an advance for this PPM?</legend>
+                      <Field
+                        as={Radio}
+                        id="yes-has-received-advance"
+                        data-testid="yes-has-received-advance"
+                        label="Yes"
+                        name="hasReceivedAdvance"
+                        value="true"
+                        checked={values.hasReceivedAdvance === 'true'}
+                      />
+                      <Field
+                        as={Radio}
+                        id="no-has-received-advance"
+                        data-testid="no-has-received-advance"
+                        label="No"
+                        name="hasReceivedAdvance"
+                        value="false"
+                        checked={values.hasReceivedAdvance === 'false'}
+                      />
+                      <Hint className={ppmStyles.hint}>
+                        If you requested an advance but did not receive it, select No.
+                      </Hint>
+                      {values.hasReceivedAdvance === 'true' && (
+                        <MaskedTextField
+                          label="How much did you receive?"
+                          name="advanceAmountReceived"
+                          id="advanceAmountReceived"
+                          defaultValue="0"
+                          mask={Number}
+                          scale={0} // digits after point, 0 for integers
+                          signed={false} // disallow negative
+                          thousandsSeparator=","
+                          lazy={false} // immediate masking evaluation
+                          prefix="$"
+                          hintClassName={ppmStyles.innerHint}
+                        />
+                      )}
+                    </Fieldset>
+                  </FormGroup>
+                  <h2>W-2 address</h2>
+                  <p>What is the address on your W-2?</p>
+                  <AddressFields
+                    name={formFieldsName}
+                    className={styles.AddressFieldSet}
+                    labelHint="Required"
+                    locationLookup
+                    formikProps={formikProps}
+                  />
+                </SectionWrapper>
+                <div
+                  className={`${
+                    isCustomerPage ? ppmStyles.buttonContainer : `${formStyles.formActions} ${ppmStyles.buttonGroup}`
+                  }`}
                 >
-                  Save & Continue
-                </Button>
-              </div>
-            </Form>
-          </div>
-        );
-      }}
-    </Formik>
+                  <Button className={ppmStyles.backButton} type="button" onClick={onBack} secondary outline>
+                    {`${isCustomerPage ? 'Return To Homepage' : 'Cancel'}`}
+                  </Button>
+                  <Button
+                    className={ppmStyles.saveButton}
+                    type="button"
+                    onClick={handleSubmit}
+                    disabled={!isValid || isSubmitted || isSubmitting}
+                  >
+                    Save & Continue
+                  </Button>
+                </div>
+              </Form>
+            </div>
+          );
+        }}
+      </Formik>
+    </>
   );
 };
