+++ src/components/Shared/PPM/Closeout/AboutForm/AboutForm.test.jsx
@@ -81,8 +81,20 @@ describe('AboutForm component', () => {
       );
 
       await waitFor(() => {
-        expect(screen.getByRole('heading', { level: 2, name: 'Departure date' })).toBeInTheDocument();
+        expect(screen.getByText('Finish moving this PPM before you start documenting it.')).toBeInTheDocument();
       });
+      const headings = screen.getAllByRole('heading', { level: 2 });
+      expect(headings[0]).toHaveTextContent('How to complete your PPM');
+      expect(headings[1]).toHaveTextContent('About your final payment');
+
+      // renders form content
+      expect(headings[2]).toHaveTextContent('Departure date');
+      expect(headings[3]).toHaveTextContent('Locations');
+      expect(headings[4]).toHaveTextContent('Advance (AOA)');
+      expect(headings[5]).toHaveTextContent('W-2 address');
+
+      expect(screen.getByRole('heading', { level: 2, name: 'Departure date' })).toBeInTheDocument();
+
       expect(screen.getByLabelText('When did you leave your origin?')).toBeInstanceOf(HTMLInputElement);
       expect(screen.getByRole('heading', { level: 2, name: 'Locations' })).toBeInTheDocument();
 
@@ -115,8 +127,19 @@ describe('AboutForm component', () => {
       );
 
       await waitFor(() => {
-        expect(screen.getByRole('heading', { level: 2, name: 'Departure date' })).toBeInTheDocument();
+        expect(screen.getByText('Finish moving this PPM before you start documenting it.')).toBeInTheDocument();
       });
+      const headings = screen.getAllByRole('heading', { level: 2 });
+      expect(headings[0]).toHaveTextContent('How to complete your PPM');
+      expect(headings[1]).toHaveTextContent('About your final payment');
+
+      // renders form content
+      expect(headings[2]).toHaveTextContent('Departure date');
+      expect(headings[3]).toHaveTextContent('Locations');
+      expect(headings[4]).toHaveTextContent('Advance (AOA)');
+      expect(headings[5]).toHaveTextContent('W-2 address');
+
+      expect(screen.getByRole('heading', { level: 2, name: 'Departure date' })).toBeInTheDocument();
       expect(screen.getByLabelText('When did you leave your origin?')).toBeInstanceOf(HTMLInputElement);
       expect(screen.getByRole('heading', { level: 2, name: 'Locations' })).toBeInTheDocument();
