+++ src/pages/MyMove/PPM/Closeout/WeightTickets/WeightTickets.jsx
@@ -23,7 +23,6 @@ import ppmPageStyles from 'pages/MyMove/PPM/PPM.module.scss';
 import NotificationScrollToTop from 'components/NotificationScrollToTop';
 import ShipmentTag from 'components/ShipmentTag/ShipmentTag';
 import { shipmentTypes } from 'constants/shipments';
-import closingPageStyles from 'pages/MyMove/PPM/Closeout/Closeout.module.scss';
 import WeightTicketForm from 'components/Shared/PPM/Closeout/WeightTicketForm/WeightTicketForm';
 import { updateAllMoves, updateMTOShipment } from 'store/entities/actions';
 import ErrorModal from 'shared/ErrorModal/ErrorModal';
@@ -215,14 +214,6 @@ const WeightTickets = () => {
             <ShipmentTag shipmentType={shipmentTypes.PPM} />
             <h1>Weight Tickets</h1>
             {renderError()}
-            <div className={closingPageStyles['closing-section']}>
-              <p>
-                Weight tickets should include both an empty or full weight ticket for each segment or trip. If you’re
-                missing a weight ticket, you’ll be able to use a government-created spreadsheet to estimate the weight.
-              </p>
-              <p>Weight tickets must be certified, legible, and unaltered. Files must be 25MB or smaller.</p>
-              <p>You must upload at least one set of weight tickets to get paid for your PPM.</p>
-            </div>
             <WeightTicketForm
               weightTicket={currentWeightTicket}
               tripNumber={currentIndex + 1}
