+++ src/pages/Office/PPM/Closeout/About/About.jsx
@@ -1,11 +1,9 @@
 import React, { useState } from 'react';
 import { generatePath, useNavigate, useParams } from 'react-router-dom';
 import { GridContainer, Grid, Alert } from '@trussworks/react-uswds';
-import classnames from 'classnames';
 import { useQueryClient, useMutation } from '@tanstack/react-query';
 
 import ppmPageStyles from 'pages/Office/PPM/PPM.module.scss';
-import closingPageStyles from 'pages/Office/PPM/Closeout/Closeout.module.scss';
 import ShipmentTag from 'components/ShipmentTag/ShipmentTag';
 import NotificationScrollToTop from 'components/NotificationScrollToTop';
 import { shipmentTypes } from 'constants/shipments';
@@ -119,33 +117,6 @@ const About = () => {
                     {errorMessage}
                   </Alert>
                 )}
-                <div className={classnames(closingPageStyles['closing-section'], closingPageStyles['about-ppm'])}>
-                  <p>Finish moving this PPM before you start documenting it.</p>
-                  <h2>How to complete your PPM</h2>
-                  <p>To complete your PPM, you will:</p>
-                  <ul>
-                    <li>Upload weight tickets for each trip</li>
-                    <li>Upload receipts to document any expenses</li>
-                    <li>Upload receipts if you used short-term storage, so you can request reimbursement</li>
-                    <li>
-                      Upload any other documentation (such as proof of ownership for a trailer, if you used your own)
-                    </li>
-                    <li>Complete your PPM to send it to a counselor for review</li>
-                  </ul>
-                  <h2>About your final payment</h2>
-                  <p>Your final payment will be:</p>
-                  <ul>
-                    <li>based on your final incentive</li>
-                    <li>modified by expenses submitted (authorized expenses reduce your tax burden)</li>
-                    <li>minus any taxes withheld (the IRS considers your incentive to be taxable income)</li>
-                    <li>plus any reimbursements you receive</li>
-                  </ul>
-                  <p>
-                    Verified expenses reduce the taxable income you report to the IRS on form W-2. They may not be
-                    claimed again as moving expenses. Federal tax withholding will be deducted from the profit
-                    (entitlement less eligible operating expenses.)
-                  </p>
-                </div>
                 <AboutForm
                   mtoShipment={mtoShipment}
                   onSubmit={handleSubmit}
